convert CollectorProcessControl to typescript and using React functional component
